requirements.txt
psutil>=5.9.0
watchdog>=3.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
Flask>=3.0.0
flask-cors>=4.0.0
python-nmap>=0.7.1
requests>=2.31.0
plyer>=2.1.0
pyyaml>=6.0.1
rich>=13.0.0
colorama>=0.4.6
joblib>=1.3.0
